cardpick: improve readbility, consolidate prime
